INIT: systemd service file

+++ Makefile
@@ -1,36 +1,36 @@
 TARGET := nv-offsetter
+SERVICE_FILE := ${TARGET}.service
+SERVICE_DIR := /usr/lib/systemd/system
 
 CXX := g++
 CXXFLAGS := -O3 -Wall -Wpedantic
-# CU := nvcc
-# CUFLAGS := -O3 -arch=sm_86
 LDFLAGS := -lnvidia-ml -lconfuse
 
 SRC := ./src/main.cpp
 BUILD_DIR := ./build
-INSTALL_DIR := /usr/local/bin/
+INSTALL_DIR := /usr/sbin
 
-.PHONY: all
 all: ${SRC} get-nvml-header
 	@if [[ ! -d ${BUILD_DIR} ]]; then mkdir ${BUILD_DIR}; fi
-	@# ${CU} ${CUFLAGS} ${LDFLAGS} ${SRC} -o ${BUILD_DIR}/${TARGET}
 	${CXX} ${CXXFLAGS} ${LDFLAGS} ${SRC} -o ${BUILD_DIR}/${TARGET}
 
-# It might not work for you, kind of hacky
-get-nvml-header:
-	@if [[ ! -d ./include ]]; then mkdir ./include ; fi
-	cp /opt/cuda/targets/x86_64-linux/include/nvml.h ./include/
+clean-build:
+	if [[ -d ${BUILD_DIR} ]]; then rm -r ${BUILD_DIR}; fi
 
-# Yeah, it's sloppy, but what are you going to do do about it?
-install-local: all
+install: all
 	install -Dm755 ${BUILD_DIR}/${TARGET} ${INSTALL_DIR}
 
-.PHONY: test
-test: all
-	@echo "There isn't proper test yet, sorry!"
+install-systemd: install
+	install -Dm644 ${SERVICE_FILE} ${SERVICE_DIR}
+
+clean: clean-build
+	if [[ -f ${INSTALL_DIR}/${TARGET} ]]; then rm ${INSTALL_DIR}/${TARGET}; fi
+	if [[ -f ${SERVICE_DIR}/${SERVICE_FILE} ]]; then rm ${SERVICE_DIR}/${SERVICE_FILE}; fi
+
+run: all
 	${BUILD_DIR}/${TARGET}
 
-.PHONY: clean
-clean:
-	rm -r ${BUILD_DIR}
-	rm ${INSTALL_DIR}${TARGET}
+get-nvml-header:
+	@if [[ ! -d ./include ]]; then mkdir ./include ; fi
+	@# It might not work for you, kind of hacky
+	cp /opt/cuda/targets/x86_64-linux/include/nvml.h ./include/
